requirements.txt
python-magic==0.4.16
Click==7.0
